requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.37.0
pydantic==2.10.5

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# AI & RAG
openai==1.58.1
langchain-core==0.3.34
langchain-community==0.3.8
langchain-openai==0.2.14
langchain-chroma==0.1.7
chromadb==1.0.20

# Environment
python-dotenv==1.0.1

# Additional AI tools
agno==2.1.4
ddgs==9.6.0
unidecode==1.3.8
